% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp.R
\name{cfp}
\alias{cfp}
\title{Cumulative Frequency Diagram}
\usage{
cfp(d, prob = 0.95, graphtitle = "CFP")
}
\arguments{
\item{d}{vector with data}

\item{prob}{which percentile to indicate on graph, default is the
95th percentile}

\item{graphtitle}{title of the graph}
}
\value{
a ggplot2 object

This function plots a Q-Q plot but instead of theoretical quantiles
it plots against the theoretical percentile, using a normal
distribution.

By default, this function indicates the values at the 95th
percentile. Using the prob parameter, the percentile can be
changed.

This function returns a ggplot2 object with the graph. This object
can be extended using ggplot2's grammar.
}
\description{
Plots a cumulative frequency diagram, a diagram comparable with a
q-q- plot but with percentiles instead of quantiles on the x-axis.
}
