% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfitci.R
\name{logfitci}
\alias{logfitci}
\title{Calculates fitted quantile and CI}
\usage{
logfitci(x, prob = 0.95)
}
\arguments{
\item{x}{Data}

\item{prob}{quantile probability}
}
\value{
A list with the fitdist and bootdist objects, and a
data.frame with a summary of of the procedure

This functions uses the `fitdist` function of the fitdistrplus
package. A log normal distribuion is fitted on the data and the
quantile and confidence interval (CI) is calculated by
bootstrapping of the fitted fucntion.
}
\description{
To calculate the quantile and CI, a log-normal distribution is
fitted on the data and the quantile and CI for this fit is returned
}
