% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfitcicens.R
\name{logfitcicens}
\alias{logfitcicens}
\title{Calculates fitted quantiles and CI, including censored data}
\usage{
logfitcicens(x, prob = 0.95)
}
\arguments{
\item{x}{Data}

\item{prob}{quantile probability}
}
\value{
A list with the fitdistcens and bootdistcens objects, and a
data,frame with a summary of of the procedure

This functions uses the `fitdistcens` function of the fitdistrplus
package. Censored data (i.e. concentrations) are indicated with
negative values. A log normal distribuion is fitted on the data and the
quantile and confidence interval (CI) is calculated by
bootstrapping of the fitted function.
}
\description{
To calculate the quantile and CI, a log-normal distribution is
fitted on the censored data and the quantile and CI for this fit is
returned.
}
