% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbci.R
\name{pbci}
\alias{pbci}
\title{Calculates bootstrapped percentiles and conf. intervals}
\usage{
pbci(x, prob = 0.95)
}
\arguments{
\item{x}{vector with data}

\item{prob}{Probability of percentile}
}
\value{
list with boot objects and summary data.frame

This functions is a wrapper arround boot and boot.ci from the boot
package. The boot statistic is the quantile function. 

The returned list contains the boot (tag b) and boot.ci (tag ci)
objects. The data from these objects is summarised in the result
tag. This contains a data.frame with, respecively, original percentile value
(original), mean of bootstrapped results (mean), bias of
bootstrapped results towards mean (bias), standard error of
bootstrapped results (error), and upper (uci) and lower lci
confidence intervals of boot.ci. Confidence intervals are
calculated for aan alpan of 0.05.
}
\description{
This function calculates a bootstrapped percentile and its
confidence interval of a given dataset.
}
