% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qestimates.R
\name{qestimates}
\alias{qestimates}
\title{Calculates quantile statistics and CI}
\usage{
qestimates(x, p = 0.95)
}
\arguments{
\item{x}{vector with data}

\item{p}{quantile probability}
}
\value{
A list with the fitdist and bootdist objects, and a
data,frame with a summary of of the procedure

This functions uses the `fitdist` function of the fitdistrplus
package. A log normal distribuion is fitted on the data and the
quantile and confidence interval (CI) is calculated by
bootstrapping of the fitted fucntion.
}
\description{
This is a wrapper function to calculate quantile statistics and CI
using the pbci, logfitci and logfitcicens functions.
}
