% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacelt.R
\name{replacelt}
\alias{replacelt}
\title{Replace '<' with negative signs}
\usage{
replacelt(x)
}
\arguments{
\item{x}{data set}
}
\value{
Dataset where '<' is replaced with '-'

In Dutch environmental reports, chemical analyses with a result
below the limit of quanitfication is often indicated with `<LOQ`,
where LOQ is the numerical value of the limit of quantification.
This function replaces the '<' character with a minus. The LOQ
value can then be turned into a negative numerical value.

When values below the limit of quantification are negative, it is
easy to do calculations or imputations. With the function
`replaceNegative` such values can be replaced using a imputation
factor.
}
\description{
Replace '<' ('lesser than' sign) 
with '-' (negative sign) for the whole dataset 
and remove spaces while we're at it.
}
