% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNegative.R
\name{replaceNegative}
\alias{replaceNegative}
\title{Replace negative values with positive ones}
\usage{
replaceNegative(x, replaceval = NA)
}
\arguments{
\item{x}{data set}

\item{replaceval}{value used for replacement}
}
\value{
Dataset with positive values

A way to work with values below the limit of Quantification
(LOQ) is to convert these values to negative concentrations.
Since negative concentrations do not exists, changing sign makes
easy to discern them from other values. Depending on the type of
analyses, these negative values can easily imputed by a positive
value, using a factor like 0.5 or 0.7, or by other values like 0 or
NA.
}
\description{
Replaces negative values, decoding for values below LOQ, into
positive values using a factor
}
