% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmoutlier.R
\name{rmoutlier}
\alias{rmoutlier}
\title{Remove outliers using the MAD criterium}
\usage{
rmoutlier(x, b = 3, mconstant = 1.4826, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of data}

\item{b}{Factor}

\item{mconstant}{MAD constant}

\item{na.rm}{remove NA's}
}
\description{
This function removes outliers using the MAD criterium according to
Leys e.a. (2013).
}
